+++ src/main/java/core/utils/Range.java
@@ -83,7 +83,7 @@ public class Range implements Cloneable, Clusterable {
             case LONG:
                 return (Long) this.high - (Long) this.low;
             case DOUBLE:
-                return (Float) this.high - (Float) this.low;
+                return (Double) this.high - (Double) this.low;
             case DATE:
                 // not quite accurate, but should be fine for estimation purposes
                 SimpleDate low = (SimpleDate) this.low;
@@ -208,8 +208,8 @@ public class Range implements Cloneable, Clusterable {
                 this.high = (long) ((Long) this.high + length * percentage);
                 break;
             case DOUBLE:
-                this.low = (float) ((Float) this.low - length * percentage);
-                this.high = (float) ((Float) this.high + length * percentage);
+                this.low = (double) ((Double) this.low - length * percentage);
+                this.high = (double) ((Double) this.high + length * percentage);
                 break;
             default:
                 throw new RuntimeException(this.type + " not implemented for range");
@@ -232,8 +232,8 @@ public class Range implements Cloneable, Clusterable {
                     splitHigh = (long) (((Long) low) + splitLength * (i + 1));
                     break;
                 case DOUBLE:
-                    splitLow = (float) (((Float) low) + splitLength * i);
-                    splitHigh = (float) (((Float) low) + splitLength * (i + 1));
+                    splitLow = (double) (((Double) low) + splitLength * i);
+                    splitHigh = (double) (((Double) low) + splitLength * (i + 1));
                     break;
                 default:
                     throw new RuntimeException("can't split non-numeric ranges");
